src/requirements.txt
azure.core
azure-ai-formrecognizer
streamlit
azure-storage-blob
python-dotenv
